fix: refactor AddQuery form structure and logic separation
